Snapshot of generated JSON from output of the dan sheet matching process.
There is one file per tab on the dan sheet.

These files are used as a local cache to iterate quickly on changes,
but also contain the reference output (the matched id for each lock).

The lock match quality is high, but no attention has yet been paid
to the quality of the urls.