Add a Scorecard option under the user menu.
There is a new route that renders a ScorecardRoute component,
which subsequently renders the Scorecard. The scorecard replaces
the read-only aspects of the current dan sheets, complete with
BB count and dan point calculations, incorporating the sameline,
upgrade, and modifier logic.

Currently the Scorecard is populated with data from the dancache.
It is also currently read-only.